remap tag with api.qdrant.tech
